requirements.txt
selenium
waitress
flask
Flask-Cors
matplotlib
torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
lancedb
transformers
pillow
requests
pandas
openai-clip